9624: Move ConditionSetModel.js -> ConditionSetFormField/model.js
