+++ app/javascript/components/DisplayLogicFormField/component.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { observer, inject, Provider } from 'mobx-react';
 
-import { provideConditionSetStore } from '../ConditionSetModel/utils';
+import { provideConditionSetStore } from '../ConditionSetFormField/utils';
 import ConditionSetFormField from '../ConditionSetFormField/component';
 import ErrorBoundary from '../ErrorBoundary/component';
